requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
youtube-transcript-api==0.6.2
deep-translator==1.11.4
pydantic==2.7.4
python-dotenv==1.0.0
requests==2.31.0